Flytter mapping fra OppgaveSpørring til JPQL fra OppgaveRepository til egen mapper. Refaktorerer noe for å sette parametere i samme metode som JPQL utledes.
